requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.8.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
httpx==0.27.2
sgp4==2.23
pymap3d==3.0.1
numpy==1.26.4
apscheduler==3.10.4
pytest==8.3.4
